requirements.txt
asgiref==3.11.0
attrs==25.4.0
cffi==2.0.0
colorama==0.4.6
cryptography==46.0.3
Django==5.2.8
django-debug-toolbar==6.2.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
exceptiongroup==1.3.1
filetype==1.2.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
pluggy==1.6.0
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-django==4.12.0
python-dotenv==1.2.1
python-magic==0.4.27
python-magic-bin==0.4.14
PyYAML==6.0.3
referencing==0.37.0
rest-framework-simplejwt==0.0.2
rpds-py==0.30.0
sqlparse==0.5.3
tomli==2.4.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
